requirements.txt
pandas
numpy
scikit-learn
pymongo[srv]
openmeteo-requests
requests-cache
retry-requests
python-dotenv
xgboost
lightgbm
mlflow
dagshub
openmeteo_requests